[#12] Services::ResolverService: extend resolve to merge add ids
Look up additional identifiers, such as LEIs, using the existing
repository, and add them to the resolved response.

It was thought that LEIs could be added to register-sources-oc
`Services::CompanyService` `get_company`, with the actual functionality
implemented by `BulkDataCompanyService` and `OcApiCompanyService`
`get_company`. Although this would work fine for `BulkDataCompanyService`,
it turns out that OpenCorporates API does not return LEIs if the
`sparse: true` parameter is used. Sparse is the current default, and
although overridden in register-v2 (and register), register-transformer-*
uses the default sparse. Changing this would be problematic, because
non-sparse API calls return far more data: for 1 company examined, 17x
more data!

After a discussion with @spacesnottabs, this alternative approach was
selected. This will avoid the sparse API calls issue, and will have the
result that regardless of whether `BulkDataCompanyService` or
`OcApiCompanyService` is used to service the request, bulk data will be
used to supply LEIs, and the OC API will never supply this information.
This would appear to satisfy the current objectives.