[#222] fix Repositories jurisdiction code, company number case
Similar to fixes to standardise OpenCorporates ID case for company
number and jurisdiction code, fix all OpenCorporates repositories to
force jurisdiction code to be lowercase, and company number to be
uppercase, when searching in Elasticsearch. Otherwise, results come back
empty.

Also fix tests to stub company number as a string, not an integer, since
this is in line with AddId, AltName, Company structs.

I'm not entirely comfortable with these changes, not because I think
that they're incorrect, but because I'm concerned about the state of
case-handling throughout Register & friends, and am wondering why
inconsistencies are occurring in the first place. However, the
OpenCorporates data is pretty clear, at least with regards to the
monthly bulk data imports, so forcing the case should at least help
results to be found in Elasticsearch. (OpenCorporates API is another
matter, and hasn't been considered for this change.)