[#216] remove AddIdRepository each_lei
This was added during previous migrations, and is not being actively
used. Remove to reduce distance between repositories, which otherwise
share a lot of code. If needed in future, it can be restored somewhere
or reimplemented.

Also remove `search_scroll` and `wrap_hit`, which are no longer used
when this is removed.